Rename `conf` to `param`.
