+++ src/runtime.rs
@@ -134,7 +134,7 @@ pub fn startup_tls_info() -> StartupTlsInfo {
 #[cfg(any(target_os = "android", target_os = "linux"))]
 #[inline]
 pub fn exe_phdrs() -> (*const c_void, usize) {
-    imp::conf::auxv::exe_phdrs()
+    imp::param::auxv::exe_phdrs()
 }
 
 #[cfg(linux_raw)]
