+++ tests/param/main.rs
@@ -1,6 +1,6 @@
-//! Tests for [`rustix::conf`].
+//! Tests for [`rustix::param`].
 
-#![cfg(feature = "conf")]
+#![cfg(feature = "param")]
 #![cfg(not(windows))]
 #![cfg_attr(target_os = "wasi", feature(wasi_ext))]
 #![cfg_attr(io_lifetimes_use_std, feature(io_safety))]
